+++ Public/client.js
@@ -1,6 +1,7 @@
 const socket = io(); // Connect to the server
 document.addEventListener('DOMContentLoaded', function() {
     populateThemeDropdown();
+    loadTheme(getCookie("themeName"));
     let username = prompt("Please enter your username:"); //Prompt for username before loading page content
     while (username === null || username.trim() === "" || username.length >=20 || username.includes("\\")) { //check valid username
         if(username.length >=20) {
@@ -65,10 +66,10 @@ document.addEventListener('DOMContentLoaded', function() {
     const themeDropdown = document.getElementById("theme-select");
     themeDropdown.onchange = function() {
         loadTheme(this.value)
+        setCookie("themeName", this.value);
     }
     vineBoom.volume = 1;
     submitButton.disabled = true; //disable submit button by default
-
 
     wagerRight.addEventListener("click", function(e) {
         e.preventDefault()
@@ -642,9 +643,11 @@ document.addEventListener('DOMContentLoaded', function() {
                         lightOptions.appendChild(newOption);
                     }
                 }
+                if(getCookie("themeName")) {
+                    themeDropdown.value = getCookie("themeName");
+                }
             });
     }
-
 
     function loadTheme(themeName) {
         fetch('./themes.json')
@@ -661,3 +664,23 @@ document.addEventListener('DOMContentLoaded', function() {
             .catch(error => console.error('Error loading the themes:', error));
     }
 });
+
+function setCookie(cname, cvalue) {
+    document.cookie = cname + "=" + cvalue + ";";
+}
+
+function getCookie(cname) {
+    let name = cname + "=";
+    let decodedCookie = decodeURIComponent(document.cookie);
+    let ca = decodedCookie.split(';');
+    for(let i = 0; i <ca.length; i++) {
+        let c = ca[i];
+        while (c.charAt(0) === ' ') {
+            c = c.substring(1);
+        }
+        if (c.indexOf(name) === 0) {
+            return c.substring(name.length, c.length);
+        }
+    }
+    return "";
+}